requirements-dev.txt
# Ferramentas de análise estática
pre-commit==3.5.0
ruff==0.1.6
mypy==1.5.1
bandit==1.7.5
pylint==3.0.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
types-requests
types-Flask
types-SQLAlchemy
types-PyJWT
types-pytz
